+++ src/main/java/Vic/exceptions/EmptyContentException.java
@@ -1,4 +1,4 @@
-package main.java.exceptions;
+package main.java.Vic.exceptions;
 
 public class EmptyContentException extends Exception {
     public EmptyContentException() {
