+++ src/main/java/Vic/exceptions/UnknownCommandException.java
@@ -1,4 +1,4 @@
-package main.java.exceptions;
+package main.java.Vic.exceptions;
 
 public class UnknownCommandException extends Exception {
     public UnknownCommandException() {
