+++ src/main/java/Vic/exceptions/VicException.java
@@ -1,4 +1,4 @@
-package main.java.exceptions;
+package main.java.Vic.exceptions;
 
 public class VicException extends Exception {
     public VicException() {
