+++ src/main/java/Vic/parser/Parser.java
@@ -1,28 +1,14 @@
-package main.java;
+package main.java.Vic.parser;
 
-import main.java.enums.FileCodes;
-import main.java.exceptions.*;
-import main.java.tasks.Deadline;
-import main.java.tasks.Event;
-import main.java.tasks.Task;
-import main.java.tasks.ToDo;
-import main.java.enums.Command;
+import main.java.Vic.tasks.TaskList;
+import main.java.Vic.ui.Ui;
+import main.java.Vic.exceptions.*;
 
-import java.io.FileWriter;
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.IOException;
-import java.io.FileReader;
-import java.io.BufferedReader;
 import java.time.Month;
 import java.time.Year;
-import java.util.*;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
 
 public class Parser {
     private static final DateTimeFormatter dateTimeFormatter = DateTimeFormatter.ofPattern("d/M/yyyy HHmm");
