+++ src/main/java/Vic/tasks/Deadline.java
@@ -1,4 +1,4 @@
-package main.java.tasks;
+package main.java.Vic.tasks;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
