+++ src/main/java/Vic/tasks/Task.java
@@ -1,4 +1,4 @@
-package main.java.tasks;
+package main.java.Vic.tasks;
 
 public class Task {
     protected String description;
